comedy_set_analysis/requirements.txt
openai==0.28
pydantic
python-dotenv
agency_swarm